+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.16.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- ensure `pyaud readme-help` results are the same between versions
 
 [0.16.0](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.16.0) - 2023-08-11
 ------------------------------------------------------------------------